update configurations for phosphor-led-manager
phosphor-led-manager no longer assumes any default led priority.

The behavior was changed in

https://github.com/openbmc/phosphor-led-manager/commit/5a9f651949c133ed0868d72734c4ac82c75a7656
and
https://github.com/openbmc/phosphor-led-manager/commit/55badf7979f57d5f9ed2db726eebf35e84d0dac5

The led priority needs to be configured, so updating all configurations
here.

Before, led-manager assumed a default led priority of 'Blink'.

So this patch is just making that explicit in the existing
configurations. There should be no behavior change from this patch.

Change-Id: I74aabb567ccf6682d56877f8bb42e636fc1746f6
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>