meta-facebook: minerva: Add blacklist config in entity-manager
Due to the CPLD on the fan board change to the XO5 starting from the DVT stage, we changed the update mechanism on it and we found that it caused the i2c bus to hang by FRUDevice initial process. So we add the i2c address of the CPLD update in the blacklist configuration to avoid this problem.

Change-Id: I55cca8ecfdc9fb84632a6cb3c2c30707ac2fef92
Signed-off-by: Yang Chen <yang.chen@quantatw.com>