meta-google: gbmc-ncsi-config: Don't override DHCP
We don't want the address and hostname presented by the NCSI hack to
override a potentially different one from DHCP. All of our platforms up
to this point used the same values for NCSI RA and DHCP, but new
platforms will not. They will expect their configs to always match the
DHCP provided values. No one is depending primarily on the NCSI RA hack
anymore.

Tested: Ran on the original smartnic platform and on a new platform
using OOB to verify that the backup addressing still works.

Change-Id: I812f5b8311158408e27127534bea4f7c4e427279
Signed-off-by: William A. Kennington III <wak@google.com>