meta-google: gbmc-bridge: Gw SRC non-FD addr support
In the past, our addressing schemes would always have the BMC at a fixed
"<pfx>:fdXX::" address in the machine. Now that we have OOB BMC access,
we need to refine the hierarchy of possible source addresses and make it
fix when the BMC has its own `<pfx>::/64`.

Tested: Ran on a classic set up machine and a machine with its own
prefix, verified that addresses fall back and forth as expected.

Change-Id: I3f86f18bbdfbea43486f154b15926b9e95a7c30b
Signed-off-by: William A. Kennington III <wak@google.com>