meta-google: gbmc-bridge: GW SRC latch gateway status
We don't want route changes to cause the networkd reloads for flapping
default gateway status. We can assume that once a node has a default
route outside the machine it will always be able to act as a router
until powercycle (we don't dynamically reconfigure network uplinks).

Tested: Subordinate nodes still do not become routes and nodes with
uplinks are still correctly identified as routers. Fewer reloads are
shown in the logs.

Change-Id: I170add23a6a24b3e184809ca24d8d1920b42d266
Signed-off-by: William A. Kennington III <wak@google.com>