meta-facebook: harma: Add power card enable check
When the CMM turns off the MB power,
it can cause the SGPIO value to change from low to high,
which can lead to the BMC misjudging the power status.
To prevent this, we will add a power card enable pin to check the status.
This pin is a direct GPIO and is pulled up by standby power.

Change-Id: Iea1eba6a591bc45d2a283720b51b7512679904d2
Signed-off-by: Peter Yin <peter.yin@quantatw.com>