meta-facebook: yosemite4: Set NIC MTU to ETH_MIN_MTU
Set the MTU size to ETH_MIN_MTU within the specified range.
According to the MCTP specification, the baseline MTU size for MCTP is 64 bytes. However, the minimum size for the Ethernet driver is set to 68
bytes. Therefore, set the NIC MTU size to 68 bytes to support NIC
out-of-band updates via MCTP protocol.

Change-Id: I491157abe979d40688f10fb1759c14034db518ea
Signed-off-by: Ian Chien <ianchien.wiwynn@gmail.com>