meta-ampere: mtjefferson: support Host consoles logger
This commit supports phosphor-hostlogger in STREAM mode, to forward
the byte stream from obmc-console to rsyslog via the imuxsock module.
The logs can be obtained in /var/log/. The logs are rotated by
logrotate. The size threshold of cpu console is 256KB and the others
are 64KB. Each log is polled every 1 minute to check if its file size
exceeds the threshold to rotate it to the new file. Each console can
be saved in up to 2 log files.

Tested (CPU console):
1. Connect to CPU SOL console
$ ssh root@<BMC_IP> -p 2200
2. Execute commands in the CPU SOL console
=> The console output on CPU SOL console will all be logged to
/var/log/obmc-console-cpu.log
=> The log will be rotated to /var/log/obmc-console-cpu.log.1 if the
original log file size exceeds 256KB every 1min poll

Signed-off-by: Hieu Huynh <hieuh@os.amperecomputing.com>
Change-Id: I2c22415bc2928732e7d7733e89fa2eb8faa13b9d