meta-google: gbmc-bridge: Run multiple DHCP clientts
We can't migrate the DHCP client off of the gbmcbr interface when a
smartnic is present if we potentially want to allow for OOB DHCP traffic
too. Instead, we should run a process specifically for the smartnic WA
if needed and keep the primary process. Allow both to execute but only
allow one to proceed with DHCP via a filesystem lock.

Tested: Ran on a machine using an OOB installation process to verify
that DHCP is still working via the expected process. Ran on a machine
that requires the smartnic WA instance of DHCP and verified it works as
expected still too.

Change-Id: Iea9c05ab99054a7066092908d81d29fbd603d64f
Signed-off-by: William A. Kennington III <wak@google.com>