entity-manager: devicetree-vpd pkg filenames fixed
Devicetree-vpd daemon in Entity-Manager was updated to remove the 'd'
at the end of the executable and service files names in a recent
revision.

The daemon and service file references now reflect the new naming.

TESTED:
Builds are able to complete even with the 'dts-vpd' PACKAGECONFIG
option enabled in the Entity-Manager recipe.

Without this fix, the builds with 'dts-vpd' enabled would stop with
an error about an unrecognized service file reference.

Change-Id: I3e010f561233407ccd94dfc6ae5fdf368919c600
Signed-off-by: Chris Sides <Christopher.Sides@hpe.com>