meta-ampere: mtmitchell: Fan config selection for 1P/2P
There was no way for phosphor-fan to know if it host boots with 1P or
2P, so when it detects absence of any of the thermal sensor used for
thermal control on Socket 1, it will consider as a sensor failure and
set fan PWM to 100%. However, the effort to support GPIO presence
detection in phosphor-fan just to decide whether to put one sensor
into consideration is big.

There's a need to support 2 different config set for mtmitchell 1P
and 2P

Their common config files will be placed in
/usr/share/phosphor-fan-presence/<daemon>/mtmitchell/ together with
the 2 different files groups_1p.json and groups_2p.json in
/usr/share/phosphor-fan-presence/control/mtmitchell/. These two files
contains the thermal sensors needed for thermal control in 1P and 2P
system which is the only different point between 1P and 2P. The file
to be loaded which is groups.json will have the same content as
groups_2p.json at build time.

phosphor-fan will read the PrettyName property under the
xyz.openbmc_projects.Inventory.Item interface to look for the
sub-folder if any config can't be found in the default location.

The property is configured in Entity Manager config so it can be
probed together with the Motherboard with the value "mtmitchell".
When BMC boots and ampere-platform-init service detects CPU1
presence, it will copy groups_2p.json into groups.json. Otherwise,
it will copy groups_1p.json into groups.json.

That way, phosphor-fan can used a different set of thermal sensors
for 1P.

Tested:
BMC boots on 1P system, in normal condition, fans are not driven to
the max speed.

Signed-off-by: Chau Ly <chaul@amperecomputing.com>
Change-Id: If67efc4792b3d3295478a892ab73bdd9f3a41fa3