meta-facebook: catalina: phosphor-gpio-monitor: add support for both evt and dvt system
Due to the linename mapping changes a lot between evt and dvt, we need
to seperate the gpio monitor and presence json file to avoid unexpected
behavior or event logs.

Add a script to check hardware revision and then create a softlink to
link to correct json file for services.

Change-Id: I75093d7e4788c1daff75c80d7a39fe666e427577
Signed-off-by: Potin Lai <potin.lai@quantatw.com>