meta-facebook: harma: reload fan control
Add a fan reload script to automatically reload the fan
when it is unplugged and a new fan is plugged in.

For example: when we remove a fan in a sensor fail status and then plug
it back in, the fan speed will remain at 50% due to the CPLD default
setting. We reload the fan service because the PID controller will not
continuously send the same FAN PWM signal.

Change-Id: Idab98f4b23f307a7cda9bcd9355adf8e42932ea6
Signed-off-by: Peter Yin <peter.yin@quantatw.com>