meta-facebook: minerva: revise fan status monitor
Due to the fan-related sensor name change starting from the DVT stage
and exporting the fan presence to the inventory, change the fan status
monitor to match the change.

Also, add the functional property check to cover the case that the FCB
AC is off but the presence is true by hardware connecting.

Change-Id: I0f6d1757ed78db2c7bad26d42c7106eff52447ef
Signed-off-by: Yang Chen <yang.chen@quantatw.com>