meta-facebook: catalina: Set PDB_LEAK_CONFIG during platform init
Due to limited space in the Power Distribution Board (PDB) CPLD for
storing additional configurations, the BMC is responsible for setting
the `PDB_LEAK_CONFIG` value during boot-up.

This implementation enables users to modify the leak configuration as
needed. The BMC saves the configuration in a file named
`catalina-platform-config.json` with the `PDB_LEAK_CONFIG` value
ranging from 0 to 3.

If the configuration file does not exist or the `PDB_LEAK_CONFIG` key is
undefined, the BMC defaults to using a leak configuration value of 0
when initializing the PDB CPLD.

Example:
```
root@bmc:~# cat /mnt/data/catalina-platform-config.json
{
  "PDB_LEAK_CONFIG": 3
}
```

This change ensures the leak configuration is properly initialized and
managed while providing flexibility for user-defined modifications.

Change-Id: I0a12117484a4c2a76d77c75f4556441ed5aca305
Signed-off-by: Potin Lai <potin.lai@quantatw.com>