meta-facebook: yosemite4: add wait-EID-remove
Description:
Add wait-until-mctp-EID-remove(180s) in chassis-powercycle
Modify the chassis_power_cycle.sh to ensure the SD/WF MCTP EIDs
to be removed after blade AC power off.

Motivation:
From blade AC power off sequence, we observed that the
rescan-fru-ocp-setting service had not completed after blade AC
powered on.

This caused the MCTP EID not to change, and consequently, no
‘add interface signal’ was sent to the pldmd service.
At this moment, the pldmd called the getSubTree method of the
object-mapper with a timeout, which caused the PLDM tree of the slot
to be moved by pldmd.

Other successful slots were able to recover because they followed the
correct sequence to remove the EID and then add it, which allowed them to
receive the 'add interface signal' necessary for the PLDM
initialization process.

The failure occurred on failed slot because the power-off sequence was
incorrect, which resulted in the loss of the 'add interface signal'
necessary for the PLDM initialization process.

Modify the chassis_power_cycle.sh to ensure the SD/WF MCTP EIDs to be
removed after blade AC power off.
Add wait time until MCTP EID to be removed after slot power off with
sleep 20s.

Test plan
Build Code -pass

Change-Id: I99398aacc8dcbcb362a698572c660721a240412b
Signed-off-by: Eric Yang <eric.yang.wiwynn@gmail.com>