meta-facebook: yosemite4: change method timeout
Description:
The default timeout of busctrl is 25s, so we add the busctrl timeout
to 120s for “busctrl RescanBus for fru rescan” due to D-bus busy

Motivation:
We found a failure is related to the timeout of “busctrl RescanBus
for fru rescan” due to D-bus busy.
That means the rescan-fru-ocp-setting service had not completed after
blade AC powered off.

Test plan:
Build code - pass

Change-Id: I7c82fd7fe39d19efad0fd51ecb37ffb4b43baed1
Signed-off-by: Eric Yang <eric.yang.wiwynn@gmail.com>