meta-facebook: harma: workaround probe device
In some cases(example: RPU not ready), the motherboard can lose power
and miss the kernel driver's probe time. To address this,
we have added a workaround to re-probe the device.

This commit will detect the power-hsd-good (standby power) signal and
perform a rescan of the fru device and rebind the driver when
the signal transitions from low to high.

Change-Id: I7b65721a4754f0a9809ae7d7938fe85d741d55bc
Signed-off-by: Peter Yin <peter.yin@quantatw.com>