meta-ibm: p10bmc: opkg: use append for PACKAGECONFIG
The base opkg has a PACKAGECONFIG of "libsolv".  This enables the
use of the external solver algorithm rather than the deprecated
internal solver.  There is a QA warning for continuing to use the
internal solver because it will be removed in a future version of
opkg.

Switch the PACKAGECONFIG to use ":append" rather than overriding
so that we get the base PACKAGECONFIG settings from the recipe,
which is currently only "libsolv".  This both removes the warning
and future proofs the machine config.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I25a665a20eca10986d001fd070089ac8a8368aee