meta-google: networking: add retries for networkd reload/reconfigure
Reload/reconfigure will be a hard failure that stops the rest of the
configuration. Sometimes systemd is already reloading and this will
fail. This retries 3 times with 3s intervals. This is needed until we
fully get rid of the reload/reconfigure.

Tested: the retry is applied when the configuration failed, no retry
happens if networkd is not activated yet.

Change-Id: I6d3a1930ce514dac7f3707917c6700c36bbf072e
Signed-off-by: Yuxiao Zhang <yuxiaozhang@google.com>