meta-openpower: fsidbg: adjust LICENSE
fsidbg is failing QA due to invalid LICENSE value ("GPL3.0").

The repository is conflicted in specification of the license.  The
LICENSE file is a GPL3 license file but the code says "version 2 ...
or any later version"[1].  Assuming the code itself is correct, this
recipe should be marked as GPL-2.0-or-later.

[1]: https://github.com/eddiejames/fsidbg/blob/14382251e92999e0646465f7c50188fce9d6f3d9/fsidbg.c#L9

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I3b6e6578338e28a6bb29cc1e73aa0af4813132c4