meta-ampere: Correct MessageID string
This commit corrects all MessageID strings used in the repository for
Redfish Event Log. The MessageID is parsed to be mapped with Redfish
Registries format in bmcweb, and it should only has 4 fields
<RegistryName.MajorVersion.MinorVersion.MessageKey>. Any redundant
string field added to that will cause a parsing failure and the log
won't appear the Redfish Event Log entries.

Signed-off-by: Chaul Ly <chaul@amperecomputing.com>
Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>
Change-Id: I4ff2cf83538b0cde6f4af001ba253825c1691332