meta-phosphor: phosphor-dbus-interfaces: use new meson.options
The recipe for phosphor-dbus-interfaces parses the meson.options
to attempt to determine which interfaces we should disable by
default (so vendor-specific interfaces are only enabled where
necessary).  The repository changed this from using the older
`meson_options.txt` to `meson.options`, so the recipe also needs to
adjust.

This should result in a flash saves on some systems.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I7d64df214bafed44662f9b357ef7bca582527513