meta-phosphor: phosphor-dbus-interfaces: fix task dependency
It was reported that the use of `deptask` is incorrect here because
that indicates that the task in all DEPENDS must be ran and doesn't
specify a dependency on tasks in this package.  Adjust to use `depends`
instead.

It was also reported there that downstream users could potentially be
patching the repository and affecting `meson.options`.  Add the
dependency on that task as well to ensure any patches are applied.

Fixes openbmc/openbmc#4012.

Change-Id: I86d9fa9aa8d046e98dbb4ddd5bcdd816538a5db2
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>