license_static: add space in append
The append call incorrectly omitted the preceding space. This was not an
issue previously because the extrausers.bbclass append ended with a
semi-colon.

That semi-colon was recently removed causing this incorrect append to
create a task that doesn't exist: [1].

This adds the space in the append to create the correct tasks.

[1]: https://github.com/openbmc/openbmc/commit/5082cc7fedfff6c28a1406c79606b09012b134bc

Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>
Change-Id: I43b327223cf4c3e81a8145c5c05d5dde14c96837