meta-ampere: remove phosphor-hwmon recipe
We tried to remove phosphor-hwmon in packagegroup-obmc-apps a long time
ago, but phosphor-hwmon is still being built because it is the default
value of VIRTUAL-RUNTIME_obmc-sensors-hwmon and we did not override
the value properly. This can be verified by checking the dependencies
graph of the final image:

  $ bitbake -g obmc-phosphor-image
  ...
  $ oe-depends-dot -k phosphor-hwmon -w ./task-depends.dot
  Because: obmc-phosphor-image packagegroup-obmc-apps
  obmc-phosphor-image -> packagegroup-obmc-apps -> phosphor-hwmon

This removes the recipe by reverting previous change and
overriding the config properly.

Tested:

  $ oe-depends-dot -k phosphor-hwmon -w ./task-depends.dot
  ERROR: Can't find key phosphor-hwmon in ./task-depends.dot

Change-Id: Id0000000703bbe24f6624275f826a7b84ecb53e6
Signed-off-by: Khang D Nguyen <khangng@os.amperecomputing.com>