meta-google: gbmc-bridge: Never terminate successful DHCP
It might be the case that one of the DHCP processes starts after the
successful terminate of another. This has happened during parallel
installs where the DHCP leases was not removed fast enough during
netboot. This can cause the RWFS to be wiped after the end of the DHCP
process, which is unexpected.

Change-Id: I37441e5c6abf69b1c001a071a2eee0966f8bf5f7
Signed-off-by: William A. Kennington III <wak@google.com>