meta-phosphor: linux: adjust kernel version name
We generate a CONFIG_LOCALVERSION in our own bbclass.  The current
implementation compares if the recipe version and the applied version
are the same and if not uses a "-dirty" indicator.  This happens
whenever someone does a downstream patch to their kernel.

"-dirty" typically means "you have uncommitted changes".  Seeing this
in your kernel version is alarming.  Change the indicator to "-patch"
to still give upstream the proper indication but to avoid the "dirty"
language.

Also disable the CONFIG_LOCALVERSION_AUTO, which would cause the
kernel build to also generate a version indicator of its own.  As a
result we end up with two different possibilities:

    - Unpatched: 6.18.8-3fcb592
    - Patched: 6.18.8-3fcb592-patch-abcdef0

For the unpatched variant, this represents the stable kernel version
that OpenBMC is currently based on, plus the specific kernel commit
from the OpenBMC tree that we are using.

For the patched variant (ie. the case where someone has downstream
patches for their kernel) this also adds a "-patch-abcdef0".  The
"abcdef0" is not necessarily useful but it gives an indication for
upstream reports that the downstream user is not using a pure tree,
but avoids the scarier "dirty" indication.  The patch commit number
isn't useful because it is as a result of bitbake applying the patches,
as patch files, to the git repository, and thus they do not reflect
any known git state outside of the bitbake unpacked source tree.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I3c9f89c646bf056c5147371260418d396c90c188