meta-facebook: catalina: Disable BIOSFRB2 watchdog at end of POST
We observed that, in some cases, the IPMI watchdog remains enabled in
the Host OS, leading to unexpected power cycles when the watchdog timer
expires. This issue occurs due to the BIOSFRB2 watchdog not properly
disabled.

To prevent unintended system resets, this commit adds a workaround to
explicitly disable the IPMI watchdog at the end of POST. This ensures
that the watchdog does not persist into the Host OS, mitigating the risk
of unexpected power cycle.

Change-Id: I3e90868ec2c2b548da76610f72b883baebd3e3a7
Signed-off-by: Potin Lai <potin.lai@quantatw.com>