led-manager: migrate configs to json
This is in preparation for [1].

There is some script [2] which looks like it's generating yaml output,
which is not considered here. It's mentioned in [3].

If you are using that script to generate, please mention.

Tested: Don't have access to any of these platforms, relying on CI
results and feedback from others.

References:

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-led-manager/+/77570
[2] https://github.com/openbmc/phosphor-mrw-tools/blob/4954d029d2d7f892976fb19d3fa9c502d64e1d1e/gen_led_groups.pl
[3] meta-phosphor/recipes-phosphor/leds/phosphor-led-manager-config-mrw-native.bb

Change-Id: I43f50ecc30f2ff3b29ca7a709a462735546bf9f9
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>