meta-facebook: ventura: add schematic init service
Due to a conflict in the power monitor IC's I2C address on the 1st and
2nd source RMC, we use the EM configuration to distinguish different
sources from the FRU field (Board Custom Data 4).
Then, use D-Bus sensors to probe the driver without properties.
Use a service to set the necessary values for the chip to ensure correct sensor readings.

Change-Id: I8373f5481aaa72652e0f8141c150b0a3225caa9f
Signed-off-by: Yang Chen <yang.chen@quantatw.com>