meta-phosphor: fan: add RDEPENDS for sensor-monitor
In order to pull in sensor-monitor previously, there was both a
PACKAGECONFIG and an image (or package group) RDEPENDS that needed
to be added.  This spread the information across multiple files and,
in the case of Minerva, was wrong, leading to a build failure on newer
Yocto subtrees.

The phosphor-fan recipe is a bit confusing because there really is no
'phosphor-fan' package but a collection of sub-packages.  It appears
that everyone that uses `phosphor-fan-sensor-monitor` also uses
`phosphor-fan-monitor`, so if we add the PACKAGECONFIG-based RDEPENDS
there it pulls in the correct dependencies.

We probably should adjust this package to not have `ALLOW_EMPTY=1` in
the future, which would likely simplify a lot of these package /
packagegroup dependencies.

Tested: Check `tmp/work/<machine>/obmc-phosphor-image/1.0/rootfs` on
yosemite4 and minerva to confirm `usr/bin/sensor-monitor` is present
on yosemite4 and missing on minerva.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I55ce61f40b86b1f1335e8440ef7d2b39f884a109