meta-phosphor: add distro feature production-level tweaks
There are some settings that are desired to be set for production-level
systems, but may be less desired in development.  Such as settings that
turn hard-errors into soft-errors.  Add a distro feature to allow
these settings to be enabled.

Add as a production-level tweak, a config for the kernel default
setting of `hung_task_panic`.  This prevents unnecessary panics
due to a busy system, usually due to poor driver implementations.

Enable the production-level tweaks on meta-facebook.

Tested:
```
root@yosemite4:~# cat /proc/sys/kernel/hung_task_panic
0
```

Change-Id: I0542547c77ddacae44d98e9ab6edc3743ff0016c
Signed-off-by: Cosmo Chou <cosmo.chou@quantatw.com>
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>