meta-google: gbmc-nic-config: Make dhcrelay dynamic
We only want to start it when the interface comes up and has addresses.
Otherwise, the daemon will fail to start in a loop and spam the journal
unnecessarily.

Tested: Ran on a machine with and without a cable to verify the dhcrelay
only runs with a cable present. Also rebooted and brought the interface
up and down 10+ times to verify daemon restart.

Change-Id: Ie42599d35106be5baeb6444dc7ddc8fa63ed1de7
Signed-off-by: William A. Kennington III <wak@google.com>