meta-facebook: minerva: add service to reconfig network setting
Meta's machines use DUID-LL as the client identifier for DHCPv6.
In Minerva CMM, the phosphor-network service retrieves the MAC address
from the FRU and generates the network config (00-bmc-eth0.network),
but it does not include the DUID-LL setting.

This temporary workaround adds a service to check and append the
DUID-LL setting to ensure the network configuration is as expected.

Change-Id: Ic3ca70e38f67bd73a61ec8ad70e46877427cfbfa
Signed-off-by: Yang Chen <yang.chen@quantatw.com>