webui-vue: Set npm cache directory
This recipe uses NPM commands but does not specify the CACHE directory.
By default, NPM stores logs and cache files in "${HOME}/.npm".

To make the Yocto/OpenEmbedded build environment more consistent and
self-contained, the CACHE directory should be set to
"${WORKDIR}/npm-cache", as specified in
poky/meta/classes-recipe/npm.bbclass configuration.

There are two ways to resolve this issue: one is to "inherit npm" to
manage the NPM package, and the other is to explicitly set the CACHE
directory. This approach adopts the simpler method of setting the NPM
cache directory.

Change-Id: I5477b28063b49e663f97522e90696dd6a91c6608
Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>