meta-facebook: yosemite4: set i3c hub register
This action is currently handled by the I3C hub driver.
However, it may result in EID loss if slots are removed.
To address this, we perform the same action in the
Early-init service by checking the presence of the slots
and writing the correct values to the register. This
approach helps resolve the issue.

Change-Id: Id1b64f954a13cb81fda855ac0611290320641e06
Signed-off-by: Sean He <sean.he.wiwynn@gmail.com>