meta-facebook: catalina: Improve host power off timing
1. Extend Graceful Shutdown Check to 30 Seconds

Observations show that the graceful-off process takes around 25 seconds.
To ensure reliable detection, the BMC check duration is extended to 30
seconds for graceful-off.

2. Increase Force-Off Power Button Press to 8 Seconds

The CPLD triggers a force-off if the power button is held for more than
6 seconds.
Extending the BMC power button press to 8 seconds ensures it exceed
CPLD’s check duration, guaranteeing a successful force-off.

Change-Id: I0d16c38ac1a8cda7733e47c9f5787fbbf328834f
Signed-off-by: Potin Lai <potin.lai@quantatw.com>