meta-facebook: catalina: Ensure graceful-off is called on host reboot
Currently, `host-graceful-poweroff@.service` conflicts with
`obmc-host-start@service`, but during a host reboot,
`obmc-host-startmin@service` is called directly. This causes
`host-graceful-poweroff@.service` to remain active when the host powers
on, preventing the graceful shutdown process from being triggered.

This fix updates `host-graceful-poweroff@.service` to conflict with
`obmc-host-startmin@service` instead, ensuring that it changes to an
inactive state correctly after host power-on so that the graceful-off
process is properly executed every time a reboot is requested.

Change-Id: I9ac3271c48323aa505c5d0e3ab45826d000e08d6
Signed-off-by: Potin Lai <potin.lai@quantatw.com>