meta-facebook: catalina: Sync host state on unexpected power changes
When the host power drops unexpectedly, `RequireHostTransition` remains
in `TransitioningToRunning`. This prevents the BMC from powering on the
host due the state already in `TransitioningToRunning`.

To address this, add a mechanism to update `RequireHostTransition` when
a power good change is detected. This ensures that power actions are not
blocked by PSM.

Change-Id: I8a49be7bee8cdafdfde16ee056a81dce8eb39f5c
Signed-off-by: Potin Lai <potin.lai@quantatw.com>