meta-facebook: minerva: modify the fan table
In the system's design of Minerva, the CMM only needs to handle the
failure case of the 24 fans with 2 zones, so change the fan duty of the normal case from 50% to 20% to avoid a higher duty value that conflicts with the duty set from other blades.

Also, change the property name in zone configuration from cycleTimeBaseMS to cycleIntervalTimeMS and extend the value to 3000 ms.

Change-Id: Ie438c60246a189c10738b2e53e2eea2aa5643d62
Signed-off-by: Yang Chen <yang.chen@quantatw.com>