meta-facebook: yosemite4: On NIC reset, wait longer before turning on hosts
With the existing script, there were instances where I could not ssh
into a machine after the power cycle service successfully completed. The
hypothesis is that, the hosts comes up too quickly. With a longer amount
of time from nic reset to host turning on, I have not been able to
reproduce the issue.

Tested: Updated the script on a YV4 and ran it locally to see if the
error could be reproduced but I was unable to.

Change-Id: Ic1e3e9a423d15eda13d4494891997156964cee51
Signed-off-by: James Athappilly <jamesatha@gmail.com>