meta-facebook: catalina: Add OSFP EEPROM driver bind service
The 24G64 EEPROM used as a 2nd source on the OSFP board exhibits
unexpected behavior when performing 1-byte address reads. This leads to
misidentification by FruDevice, which incorrectly detects it as a
1-byte address EEPROM and fails to retrieve the correct FRU content.

To address this issue, add a service that waits for host power-on and
then binds the OSFP EEPROM driver. This ensures that FruDevice
accesses the EEPROM via the proper driver, preventing incorrect reads.

Change-Id: I7d8fc49419e34173dd813b4520ed98ea90b411aa
Signed-off-by: Potin Lai <potin.lai@quantatw.com>