meta-ibm: Inhibit phosphor-led-manager-config-mrw
The LED yaml formatting is being replaced[1].
Since the HW design for witherspoon is not planned to change, checkin
the latest generated yaml file so that it can be converted into json,
instead of updating the MRW script that generates the yaml file.

Tested: Verified that the led.yaml file under
```
phosphor-led-manager/1.0+git/recipe-sysroot-native/usr/share/phosphor-led-manager/led.yaml
```
was the same with this change.

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-led-manager/+/77570

Change-Id: I51fa3ff5ccd518aa721e00d169d4276708dfe2a0
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>