meta-nvidia: meta-gb200nvl-obmc: add gpu card conf
Add a config for an add-in Nvidia GPU PCIe card - this is a GPU card
that can be plugged into a GB200NVL (or other) platform.

The config will allow for specific inclusion (or removal) of packages to
suit GPU PCIe card management. This condig can be included in build's
local.conf.

Change-Id: If138d247a36752def57508017f0d96185bfea853
Signed-off-by: Deepak Kodihalli <deepak.kodihalli.83@gmail.com>