meta-google: gbmc-bridge: Enable RA after gbmcbr is configured
There could be a race condition that RA is detected before the link
local address is set, causing configuration failure due to invalid
gateway. This will make sure that RA is send after the address

Tested: On a machine, verified that gbmc-br-ra is still running
correctly and observed that it is able to delayed after it is
configured.

Change-Id: I20c17a6f6a09335b376a4350d6f6043e25b52e7f
Signed-off-by: yuxiaozhang <yuxiaozhang@google.com>