meta-facebook: harma: fix power state abnormal log
Remove current status synchronization on GPIO change, as it may
lead to abnormal events being recorded by the PSM or GPIO service.

For example, when the power button is pressed to power off the host:

Power off:
The GPIO service updates CurrentHostState to Off, which records the
first "Host off" log. Then it synchronizes RequestedHostTransition to
Off, causing CurrentHostState to change from TransitioningToOff to Off,
which results in a second "Host off" log being recorded.

The GPIO service updates CurrentHostState to Running, and then
synchronizes RequestedHostTransition to Running. If CurrentHostState is
already Running, the PSM considers this an abnormal state. As a result,
it triggers a host reboot status, which leads to on -> off -> on
in the logs.

Change-Id: I543c2d46831a182ee1879a1e983d3ccc2509a0ae
Signed-off-by: Peter Yin <peter.yin@quantatw.com>