meta-facebook: yosemite4: set NIC version property correctly when plugged or unplugged
When the NIC is plugged or unplugged, the version property
is not set, so the version will still persist. Modify the
configure-nic-endpoint script so that when the NIC is plugged in,
the version will be set, and when it is unplugged, the version
will be set to an empty string.

Change-Id: I7bf0bf0f8426576d9756e5f3ef2280e94b985815
Signed-off-by: Sean He <sean.he.wiwynn@gmail.com>