meta-facebook: revamp dbus-sensors PACKAGECONFIG
Revamp PACKAGECONFIG so that all systems use `:append` rather than
setting the final value of the PACKAGECONFIG.  This reduces the amount
of per-system configuration we need to do.

Also introduce a variable FACEBOOK_REMOVED_DBUS_SENSORS for
PACKAGECONFIG removal so that systems can modify that variable to
override a default meta-facebook removal (for example in rare cases
when "external" needs to be leveraged).

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I0e575c3152162f597621c5458af8b72f9c3bc811